requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
python-dotenv
SQLAlchemy
Werkzeug
python-binance 
matplotlib
Pillow 
pytz 